✨Initial implementation of movie credits scroll
Inspired by https://twitter.com/shanselman/status/1357036562077241344
this is a very basic scroll of movie credits like foreground that can
be added as an OBS browser to overlay ontop of a live video. It looks
cool.

TODO is to get better fake job data by parsing common jobs out of an
imdb file.

TODO it would be very cool to interface with a meeting API to automatically
add call attendees.